separate routes to started and learn files, fix compositions and docs
