+++ docs/ui/community-docs/community-docs.composition.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
 import { CommunityDocs } from './community-docs';
 
-export const BasicCommunityDocs = () => (
-  <CommunityDocs />
-);
+export const BasicCommunityDocs = () => <CommunityDocs />;
