+++ docs/ui/community-docs/community-docs.docs.mdx
@@ -1,6 +1,6 @@
 ---
 description: 'A CommunityDocs component.'
-labels: ['label1', 'label2', 'label3']
+labels: ['docs', 'page', 'routes']
 ---
 
 import { CommunityDocs } from './community-docs';
@@ -10,13 +10,15 @@ import { CommunityDocs } from './community-docs';
 A basic div that renders some text
 
 ### Component usage
+
 ```js
-<CommunityDocs text="hello from CommunityDocs" />
+<CommunityDocs />
 ```
 
 ### Using props to customize the text
 
 Modify the text to see it change live:
+
 ```js live
-<CommunityDocs text="hello from CommunityDocs" />
+<CommunityDocs />
 ```
