+++ docs/ui/community-docs/community-docs.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
 import { Docs, DocsProps } from '@teambit/docs.ui.docs';
 // import { ContributingDocs } from '@teambit/docs.content.contributing-docs';
-import { learndDocsRoutes, gettingStartedDocsRoutes } from './docs-routes';
 import { primaryRoutes } from './primary-routes';
+import { gettingStartedDocsRoutes } from './getting-started-routes';
+import { learndDocsRoutes } from './learn-routes';
 
 export type CommunityDocsProps = {
   /**
    * base url to use for docs section.
    */
   baseUrl?: string;
-} & Omit<DocsProps, 'routes'>;
+} & Omit<DocsProps, 'gettingStartedRoutes' | 'learndRoutes'>;
 
 export function CommunityDocs({ baseUrl = '/docs', ...rest }: CommunityDocsProps) {
   return (
