+++ docs/ui/docs/docs.composition.tsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import { Docs } from './docs';
 
-export const BasicDocs = () => <Docs routes={[]} />;
+export const BasicDocs = () => <Docs gettingStartedRoutes={[]} learndRoutes={[]} />;
