+++ docs/ui/docs/docs.docs.mdx
@@ -1,6 +1,6 @@
 ---
 description: 'A Docs component.'
-labels: ['label1', 'label2', 'label3']
+labels: ['docs', 'page']
 ---
 
 import { Docs } from './docs';
@@ -10,13 +10,15 @@ import { Docs } from './docs';
 A basic div that renders some text
 
 ### Component usage
+
 ```js
-<Docs text="hello from Docs" />
+<Docs gettingStartedRoutes={[]} learndRoutes={[]} />
 ```
 
 ### Using props to customize the text
 
 Modify the text to see it change live:
+
 ```js live
-<Docs text="hello from Docs" />
+<Docs gettingStartedRoutes={[]} learndRoutes={[]} />
 ```
